Add logic for loading pages
